Minor code cleanups - but note port change for #23196 [ISC-Bugs #23470] - Modify when an ignore return macro is defined to handle unsed error return warnings for more versions of gcc. [ISC-Bugs #23196] - Modify the reply handling in the server code to send to a specified port rather than to the source port for the incoming message.  Sending to the source port was test code that should have been removed.  The previous functionality may be restored by defining REPLY_TO_SOURCE_PORT in the includes/site.h file.  We suggest you don't enable this except for testing purposes. [ISC-Bugs #22695] - Close a file descriptor in an error path. [ISC-Bugs #19368] - Tidy up variable types in validate_port.

+++ common/inet.c
@@ -4,8 +4,9 @@
    way... */
 
 /*
- * Copyright (c) 2004,2005,2007-2009 by
- *				Internet Systems Consortium, Inc. ("ISC")
+ * Copyright (c) 2011 by Internet Systems Consortium, Inc. ("ISC")
+ * Copyright (c) 2007-2009 by Internet Systems Consortium, Inc. ("ISC")
+ * Copyright (c) 2004,2005 by Internet Systems Consortium, Inc. ("ISC")
  * Copyright (c) 1995-2003 by Internet Software Consortium
  *
  * Permission to use, copy, modify, and distribute this software for any
@@ -611,9 +612,9 @@ piaddrcidr(const struct iaddr *addr, unsigned int bits) {
 
 u_int16_t
 validate_port(char *port) {
-	int local_port = 0;
-	int lower = 1;
-	int upper = 65535;
+	long local_port = 0;
+	long lower = 1;
+	long upper = 65535;
 	char *endptr;
 
 	errno = 0;
@@ -623,8 +624,8 @@ validate_port(char *port) {
 		log_fatal ("Invalid port number specification: %s", port);
 
 	if (local_port < lower || local_port > upper)
-		log_fatal("Port number specified is out of range (%d-%d).",
+		log_fatal("Port number specified is out of range (%ld-%ld).",
 			  lower, upper);
 
-	return htons(local_port);
+	return htons((u_int16_t)local_port);
 }
